requirements.txt
python-chess
pathlib
